fix(optimizer/gpu): robust first-step detection + Metal/OpenCL/Embedding guards (#719)
Second batch of #719 review findings:

- int8 first-step detection keyed on beta1 alone (biasCorrection1 <= 1-beta1) is
  ALWAYS true when beta1==0, so a beta1==0/beta2>0 config discarded v history every
  step. Require BOTH corrections to indicate step 1 across all paths: CUDA, HIP,
  OpenCL, WebGpu int8 shader, and the CPU CompressedMomentHostFallback.
- MetalBuffer CopyBytesTo/From: `offsetInBytes + count` was evaluated as int and
  could overflow past the bounds check before Marshal.Copy — evaluate in long.
- DirectGpuTensorEngine explicit IEngine.Embedding: validate indices against
  vocabSize (mirrors the public override) before the GPU lookup, so an out-of-range
  index falls back to base instead of an out-of-bounds device read.
- DirectGpuTensorEngine FlashAttentionBackward: drop a redundant second zero-fill of
  gradQ/K/V (nothing between the two fills touched the buffers).
- OpenCL adam8bit: the kernel needs a fixed 256-wide workgroup reduction; fail with a
  clear error on a device whose max workgroup size < 256 instead of an opaque
  CL_INVALID_WORK_GROUP_SIZE (clamping down would corrupt the reduction).

138 optimizer/serialization/embedding tests green.

+++ src/AiDotNet.Tensors/Engines/DirectGpu/Metal/MetalBuffer.cs
@@ -284,7 +284,9 @@ public sealed class MetalGpuBuffer : IGpuBuffer
             throw new ArgumentOutOfRangeException(nameof(offsetInBytes));
         }
 
-        if (count < 0 || offsetInBytes + count > SizeInBytes || count > destination.Length)
+        // (long) cast: offsetInBytes + count are both int, so a large offset+count would overflow to a
+        // negative int and slip past this bound before reaching Marshal.Copy. Evaluate in long.
+        if (count < 0 || (long)offsetInBytes + count > SizeInBytes || count > destination.Length)
         {
             throw new ArgumentOutOfRangeException(nameof(count));
         }
@@ -325,7 +327,8 @@ public sealed class MetalGpuBuffer : IGpuBuffer
             throw new ArgumentOutOfRangeException(nameof(offsetInBytes));
         }
 
-        if (count < 0 || offsetInBytes + count > SizeInBytes || count > source.Length)
+        // (long) cast: see CopyBytesTo — int offsetInBytes + count can overflow past this bound.
+        if (count < 0 || (long)offsetInBytes + count > SizeInBytes || count > source.Length)
         {
             throw new ArgumentOutOfRangeException(nameof(count));
         }
